backend/requirements.txt
Django>=3.2.0,<5.2.1
djangorestframework>=3.12.0,<4.0.0
djangorestframework-simplejwt>=5.2.0,<6.0.0
django-cors-headers>=3.13.0,<4.0.0
Pillow>=9.0.0,<10.0.0  # For image handling
python-magic>=0.4.27,<0.5.0  # For file type validation
pyotp>=2.8.0,<3.0.0  # For two-factor authentication
qrcode>=7.3.1,<8.0.0  # For 2FA QR codes
django-filter>=22.1,<23.0.0
drf-yasg>=1.21.0,<2.0.0  # For Swagger/OpenAPI documentation
python-dotenv>=0.21.0,<0.22.0
psycopg2-binary>=2.9.5,<3.0.0  # PostgreSQL adapter
gunicorn>=20.1.0,<21.0.0
whitenoise>=6.2.0,<7.0.0  # For serving static files
django-storages>=1.13.1,<2.0.0  # For cloud storage (optional)
boto3>=1.26.0,<2.0.0  # For AWS S3 (optional)
camunda-external-task-client-python>=4.0.0  # For Camunda BPM integration
requests>=2.28.0,<3.0.0  # Required for Camunda client